+++ HISTORY.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Fix check for default `secure` option behavior
+  * Fix `maxAge` option preventing cookie deletion
   * deps: depd@~2.0.0
     - Replace internal `eval` usage with `Function` constructor
     - Use instance methods on `process` to check for listeners